Feat: 상담 예약 페이지 API 연결 및 코드 정리
